시큐리티 설정, SiteUser 클래스 생성, id Long으로 수정

+++ src/main/java/com/ll/exam/sbb/answer/AnswerController.java
@@ -21,7 +21,7 @@ public class AnswerController {
     private final AnswerService answerService;
 
     @RequestMapping("/create/{id}")
-    public String createAnswer(Model model, @PathVariable int id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
+    public String createAnswer(Model model, @PathVariable Long id, @Valid AnswerForm answerForm, BindingResult bindingResult) {
         Question question = this.questionService.getQuestion(id);
 
         if (bindingResult.hasErrors()) {
